requirements.txt
# Deep Learning & ML
tensorflow==2.16.1
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# XAI Libraries

lime==0.2.0.1

# Backend
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Data Processing
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
